input.py (inside grin package)
Created a function get_input() inside the input.py module that reads input and converts it to grin tokens and returns a list of grin tokens. Did not make a unit test for this though, as input cannot be tested without the downloading of outside resources from Python Standard Library.